web: add StatusBarItem
