+++ web/src/components/core/StatusBar/StatusBarItem.css
@@ -1,0 +1,49 @@
+.StatusBarItem {
+  color: #fff;
+  border: 0;
+  background: transparent;
+  font-family: inherit;
+  font-size: 14px;
+  font-weight: 400;
+  display: flex;
+  height: 100%;
+  padding: 0 5px;
+  white-space: pre;
+  align-items: center;
+  text-overflow: ellipsis;
+  overflow: hidden;
+  outline-width: 0;
+  margin: 0 3px;
+}
+
+.StatusBarItem.StatusBarItem--button {
+  cursor: pointer;
+}
+
+.StatusBarItem.StatusBarItem--text {
+  user-select: none;
+  -moz-user-select: none;
+  -webkit-user-select: none;
+}
+
+.StatusBarItem__label {
+  margin-left: 0.3em;
+  font-size: 12px;
+}
+
+.StatusBarItem__icon {
+  font-size: 12px;
+}
+
+.StatusBarItem.StatusBarItem--button:not([disabled]):hover {
+  background: rgba(255, 255, 255, 0.1);
+}
+
+.StatusBarItem.StatusBarItem--button:not([disabled]):active {
+  background: rgba(255, 255, 255, 0.2);
+}
+
+.StatusBarItem.StatusBarItem--button[disabled] {
+  cursor: not-allowed;
+  opacity: 0.3;
+}
